Refactor read and write locks to minimize lock duration
Addresses issue #249

There are some lock acquisition behavior that exists in `service.rs` which leads to
non-obvious lock retention lifetimes.  The syntactic sugar pattern matching that
rust allows ends up hiding the lifetime acquisition of references to a locked
global state.  These locks end up getting held across async boundaries unnecessarily
which will only lead to lock contention when it is not needed.

Additionally, in at least one case, the lock acquisition lifetimes will result in a deadlock
as a `write_arc` is being attempted when the write lock is already being held in a
parent closure: `service.rs:630`.

In order to improve the lock contention and to avoid the deadlock potential the lock
guards have been made explicit and placed behind a closure ensuring that it will
be cleaned up as soon as the information needed from it is accessed.  Any information
outside of the lock is cloned in order to ensure no further lock is required.  The data
that has been cloned / copied out of the lock fields was already being cloned / copied
regardless.  The end result should be an improved representation of the lock guard's
lifetimes, and clarity of the closures that hold a lock guard.